event-platform/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
stripe==7.8.0
paypalrestsdk==1.13.1
python-multipart==0.0.6
qrcode[pil]==7.4.2
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
emails==0.6
jinja2==3.1.2
pandas==2.1.4
openpyxl==3.1.2
redis==5.0.1
celery==5.3.4
